locator: Keep track of the distance travelled
Previously, we would emit the LocationUpdated signal for any location
property change, disregarding the DistanceThreshold value. A change in
the "description" property would trigger the LocationUpdated signal,
even if we were well below the threshold.

This commit changes it to emit LocationUpdated only when we cross the
threshold. We keep track of the distance travelled and while it's below
the distance threshold, we update the location properties, but don't
emit the signal.

https://bugs.freedesktop.org/show_bug.cgi?id=69105